HSM and logging

+++ sm/csm/utils.c
@@ -43,7 +43,7 @@ void my_basename_no_suffix(const char *file, char*dest) {
   const char *end = strrchr(file, '.');
   if(!end || end<start) end = start + strlen(start);
 
-  strncpy(dest, start, end-start);
+  strncpy(dest, start, (size_t) (end-start));
   dest[end-file] = '\0';
 //	sm_debug("start: '%s' end: '%s' base_no_suffix: '%s'\n", start, end, dest);
 }
@@ -51,7 +51,7 @@ void my_basename_no_suffix(const char *file, char*dest) {
 void my_no_suffix(const char *file, char*dest) {
   const char *end = strrchr(file, '.');
   if(!end) end = file + strlen(file);
-  strncpy(dest, file, end-file);
+  strncpy(dest, file, (size_t) (end-file) );
   dest[end-file] = '\0';
 }
